+++ packages/tswebextension/src/lib/mv3/background/declarative-filtering-log.ts
@@ -1,4 +1,4 @@
-import { IRuleSet, SourceRuleAndFilterId } from '@adguard/tsurlfilter';
+import { IRuleSet, SourceRuleAndFilterId } from '@adguard/tsurlfilter/es/declarative-converter';
 
 /**
  * Information about applied declarative network rule.