+++ packages/tswebextension/src/lib/mv3/background/index.ts
@@ -14,7 +14,7 @@ import {
     TooManyRulesError,
     UnavailableFilterSourceError,
     UnavailableRuleSetSourceError,
-} from '@adguard/tsurlfilter';
+} from '@adguard/tsurlfilter/es/declarative-converter';
 
 export * from './app';