+++ packages/tswebextension/src/lib/mv3/background/rule-sets-loader-api.ts
@@ -11,7 +11,7 @@ import {
     RULES_COUNT_FILENAME,
     SOURCE_MAP_FILENAME_JSON,
     DeclarativeRuleValidator,
-} from '@adguard/tsurlfilter';
+} from '@adguard/tsurlfilter/es/declarative-converter';
 import { z as zod } from 'zod';
 
 /**