+++ packages/tswebextension/src/lib/mv3/background/user-rules-api.ts
@@ -3,7 +3,7 @@ import {
     Filter,
     ConversionResult,
     IFilter,
-} from '@adguard/tsurlfilter';
+} from '@adguard/tsurlfilter/es/declarative-converter';
 
 export { ConversionResult };